+++ Sources/NSFetchRequestResult+CoreData.swift
@@ -2,41 +2,68 @@
 
 import CoreData
 
-extension NSFetchRequestResult where Self: NSManagedObject {
+extension NSManagedObject {
   /// The entity name.
-  /// - Warning: The NSManagedObjectModel must be loaded or the execution will be stopped.
-  public static var entityName: String {
+  ///
+  /// Usually there is no need to override this property; once overridden be sure to ovverride it in all the subclasses (if any).
+  /// - Warning: If the `NSManagedObjectModel` is not loaded, the value will fallback to a string representation of `Self`.
+  @objc open class var entityName: String {
     if let name = entity().name {
       return name
     }
+    
+    // Fallback to string representation of Self
+    
     // https://stackoverflow.com/questions/37909392/exc-bad-access-when-calling-new-entity-method-in-ios-10-macos-sierra-core-da
     // https://stackoverflow.com/questions/43231873/nspersistentcontainer-unittests-with-ios10/43286175
     // https://www.jessesquires.com/blog/swift-coredata-and-testing/
     // https://github.com/jessesquires/rdar-19368054
-
+    
     // Returning a string representation of Self metatype doesn't work if you have a NSManagedObject subclass
-    // with a name different from the NSEntityDescription name.
-    // return String(describing: Self.self)
-
-    // see testEntityName()
-    fatalError("Have you loaded your NSManagedObjectModel yet?")
+    // with a name different from the NSEntityDescription name (In that case is ovverride this property).
+    
+    return String(describing: Self.self)
   }
+}
 
+//extension NSFetchRequestResult where Self: NSManagedObject {
+//  /// The entity name.
+//  /// - Warning: The `NSManagedObjectModel` must be loaded or the execution will be stopped.
+//  public static var _entityName: String {
+//    if let name = entity().name {
+//      return name
+//    }
+//    
+//    // https://stackoverflow.com/questions/37909392/exc-bad-access-when-calling-new-entity-method-in-ios-10-macos-sierra-core-da
+//    // https://stackoverflow.com/questions/43231873/nspersistentcontainer-unittests-with-ios10/43286175
+//    // https://www.jessesquires.com/blog/swift-coredata-and-testing/
+//    // https://github.com/jessesquires/rdar-19368054
+//    
+//    // Returning a string representation of Self metatype doesn't work if you have a NSManagedObject subclass
+//    // with a name different from the NSEntityDescription name.
+//    // return String(describing: Self.self)
+//    
+//    fatalError("Have you loaded your NSManagedObjectModel yet?")
+//  }
+//}
+
+extension NSFetchRequestResult where Self: NSManagedObject {
+  
   // MARK: - Fetch
-
+  
   /// Returns a new fetch request initialized with the entity represented by this subclass (`self`).
   /// - Warning: This fetch request is created with a string name (`entityName`), and cannot respond to -entity until used by an NSManagedObjectContex.
   public static func newFetchRequest() -> NSFetchRequest<Self> {
     NSFetchRequest<Self>(entityName: entityName)
   }
-
+  
   /// - Returns: an object for a specified `id` even if the object needs to be fetched.
   /// If the object is not registered in the context, it may be fetched or returned as a fault.
   /// If use existingObject(with:) if you don't want a faulted object.
   public static func object(with id: NSManagedObjectID, in context: NSManagedObjectContext) -> Self? {
     context.object(with: id) as? Self
   }
-
+  
   /// - Returns: the object for the specified ID or nil if the object does not exist.
   /// If there is a managed object with the given ID already registered in the context, that object is returned directly; otherwise the corresponding object is faulted into the context.
   /// This method might perform I/O if the data is uncached.
@@ -44,7 +71,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   public static func existingObject(with id: NSManagedObjectID, in context: NSManagedObjectContext) throws -> Self? {
     try context.existingObject(with: id) as? Self
   }
-
+  
   /// Performs a configurable fetch request in a context.
   /// - Note: It always accesses the underlying persistent stores to retrieve the latest results.
   /// - Attention: Core Data makes heavy use of Futures, especially for relationship values.
@@ -56,9 +83,8 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ///   - configuration: Configuration closure applied **only** before fetching.
   /// - Throws: It throws an error in cases of failure.
   /// - Returns: An array of objects meeting the criteria specified by request fetched from *the receiver* and from *the persistent stores* associated with the receiver’s persistent store coordinator.
-  public static func fetchObjects(
-    in context: NSManagedObjectContext, with configuration: (NSFetchRequest<Self>) -> Void = { _ in }
-  ) throws -> [Self] {
+  public static func fetchObjects(in context: NSManagedObjectContext,
+                                  with configuration: (NSFetchRequest<Self>) -> Void = { _ in }) throws -> [Self] {
     // Check the Discussion paragraph for the fetch(_:) documentation:
     // https://developer.apple.com/documentation/coredata/nsmanagedobjectcontext/1506672-fetch
     // When you execute an instance of NSFetchRequest, it always accesses the underlying persistent stores to retrieve the latest results.
@@ -70,7 +96,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
       "This method requires a NSFetchRequest with resultType of .managedObjectResultType.")
     return try context.fetch(request)
   }
-
+  
   /// Performs a configurable fetch request in a context.
   /// - Note: When fetching data from Core Data, you don’t always know how many values you’ll be getting back.
   /// Core Data solves this problem by using a subclass of `NSArray` that will dynamically pull in data from the underlying store on demand.
@@ -87,15 +113,15 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ) throws -> NSArray {
     // Check the Discussion paragraph for the fetch(_:) documentation:
     // https://developer.apple.com/documentation/coredata/nsmanagedobjectcontext/1506672-fetch
-
+    
     // When you execute an instance of NSFetchRequest, it always accesses the underlying persistent stores to retrieve the latest results.
     // https://developer.apple.com/documentation/coredata/nsfetchrequest
     let request = NSFetchRequest<Self>(entityName: entityName)
     configuration(request)
-
+    
     return try context.fetchNSArray(request)
   }
-
+  
   /// Fetches all the `NSManagedObjectID` for a given predicate.
   /// - Note: it always accesses the underlying persistent stores to retrieve the latest results.
   ///
@@ -128,12 +154,12 @@ extension NSFetchRequestResult where Self: NSManagedObject {
     request.includesSubentities = includingSubentities
     request.predicate = predicate
     request.affectedStores = affectedStores
-
+    
     return try context.fetch(request)
   }
-
+  
   // MARK: - First
-
+  
   /// Fetches an object matching the given predicate.
   /// - Note: it always accesses the underlying persistent stores to retrieve the latest results.
   ///
@@ -158,9 +184,9 @@ extension NSFetchRequestResult where Self: NSManagedObject {
       request.fetchLimit = 1
     }.first
   }
-
+  
   // MARK: - Unique
-
+  
   /// Executes a fetch request where **at most** a single object is expected as result; if more than one object are fetched, a fatal error will occour.
   /// - Note: To guarantee uniqueness the fetch accesses the underlying persistent stores to retrieve the latest results and, also, matches against currently
   /// unsaved changes in the managed object context.
@@ -173,7 +199,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   public static func fetchUniqueObject(
     in context: NSManagedObjectContext, where predicate: NSPredicate, affectedStores: [NSPersistentStore]? = nil
   )
-    throws -> Self?
+  throws -> Self?
   {
     let result = try fetchObjects(in: context) { request in
       request.predicate = predicate
@@ -181,7 +207,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
       request.affectedStores = affectedStores
       request.fetchLimit = 2
     }
-
+    
     switch result.count {
     case 0:
       return nil
@@ -191,9 +217,9 @@ extension NSFetchRequestResult where Self: NSManagedObject {
       fatalError("Returned multiple objects, expected max 1.")
     }
   }
-
+  
   // MARK: - Delete
-
+  
   /// Specifies the objects (matching a given `predicate`) that should be removed from its persistent store when changes are committed.
   /// If objects have not yet been saved to a persistent store, they are simply removed from the context.
   /// If the dataset to delete is very large, use the `limit` value to decide the number of objects to be deleted otherwise the operation could last an unbounded amount time.
@@ -221,7 +247,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
       }.lazy.forEach(context.delete(_:))
     }
   }
-
+  
   /// Removes all entities from within the specified `NSManagedObjectContext` excluding a given list of entities.
   ///
   /// - Parameters:
@@ -236,9 +262,9 @@ extension NSFetchRequestResult where Self: NSManagedObject {
     let predicate = NSPredicate(format: "NOT (self IN %@)", objects)
     try delete(in: context, includingSubentities: true, where: predicate, affectedStores: affectedStores)
   }
-
+  
   // MARK: - Count
-
+  
   /// Counts the results of a configurable fetch request in a context.
   /// - Throws: It throws an error in cases of failure.
   public static func count(
@@ -246,16 +272,16 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ) throws -> Int {
     let request = newFetchRequest()
     configuration(request)
-
+    
     let result = try context.count(for: request)
     // result is equal to NSNotFound if an error occurs (an exception is expected to be thrown)
     guard result != NSNotFound else { return 0 }
-
+    
     return result
   }
-
+  
   // MARK: - Materialized Object
-
+  
   /// Iterates over the context’s registeredObjects set (which contains all managed objects the context currently knows about) until it finds one that is not a fault matching for a given predicate.
   /// Faulted objects are not considered to prevent Core Data to make a round trip to the persistent store.
   ///
@@ -266,12 +292,12 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   public static func materializedObject(in context: NSManagedObjectContext, where predicate: NSPredicate) -> Self? {
     for object in context.registeredObjects where !object.isFault {
       guard let result = object as? Self, predicate.evaluate(with: result) else { continue }
-
+      
       return result
     }
     return nil
   }
-
+  
   /// Iterates over the context’s registeredObjects set (which contains all managed objects the context currently knows about) until it finds
   /// all the objects that aren't a fault matching for a given predicate.
   /// Faulted objects are not considered to prevent Core Data to make a round trip to the persistent store.
@@ -300,14 +326,14 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   /// - Returns: The result returned when executing a batch update request.
   /// - Note: A batch delete can **only** be done on a SQLite store.
   public static func batchUpdate(using context: NSManagedObjectContext, configuration: (NSBatchUpdateRequest) -> Void)
-    throws -> NSBatchUpdateResult
+  throws -> NSBatchUpdateResult
   {
     let batchRequest = NSBatchUpdateRequest(entityName: entityName)
     configuration(batchRequest)
     // swiftlint:disable:next force_cast
     return try context.execute(batchRequest) as! NSBatchUpdateResult
   }
-
+  
   /// Executes a batch delete on the context's persistent store coordinator.
   /// - Parameters:
   ///   - context: The context whose the persistent store coordinator will be used to execute the batch delete.
@@ -333,16 +359,16 @@ extension NSFetchRequestResult where Self: NSManagedObject {
     let request = NSFetchRequest<Self>(entityName: entityName)
     request.predicate = predicate
     request.includesSubentities = includesSubentities
-
+    
     // swiftlint:disable:next force_cast
     let batchRequest = NSBatchDeleteRequest(fetchRequest: request as! NSFetchRequest<NSFetchRequestResult>)
     batchRequest.resultType = resultType
     batchRequest.affectedStores = affectedStores
-
+    
     // swiftlint:disable:next force_cast
     return try context.execute(batchRequest) as! NSBatchDeleteResult
   }
-
+  
   /// Executes a batch insert on the context's persistent store coordinator.
   /// - Parameters:
   ///   - context: The context whose the persistent store coordinator will be used to execute the batch insert.
@@ -360,11 +386,11 @@ extension NSFetchRequestResult where Self: NSManagedObject {
     let batchRequest = NSBatchInsertRequest(entityName: entityName, objects: objects)
     batchRequest.resultType = resultType
     batchRequest.affectedStores = affectedStores
-
+    
     // swiftlint:disable:next force_cast
     return try context.execute(batchRequest) as! NSBatchInsertResult
   }
-
+  
   /// Executes a batch insert on the context's persistent store coordinator.
   /// Doing a batch insert with this method is more memory efficient than the standard batch insert where all the items are passed alltogether.
   /// - Parameters:
@@ -381,11 +407,11 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ) throws -> NSBatchInsertResult {
     let batchRequest = NSBatchInsertRequest(entityName: entityName, dictionaryHandler: handler)
     batchRequest.resultType = resultType
-
+    
     // swiftlint:disable:next force_cast
     return try context.execute(batchRequest) as! NSBatchInsertResult
   }
-
+  
   /// Executes a batch insert on the context's persistent store coordinator.
   /// Doing a batch insert with this method is more memory efficient than the standard batch insert where all the items are passed alltogether.
   /// - Parameters:
@@ -407,7 +433,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
         handler(object as! Self)
       })
     batchRequest.resultType = resultType
-
+    
     // swiftlint:disable:next force_cast
     return try context.execute(batchRequest) as! NSBatchInsertResult
   }
@@ -445,7 +471,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ) throws -> NSAsynchronousFetchResult<Self> {
     let request = Self.newFetchRequest()
     configuration(request)
-
+    
     let asynchronousRequest = NSAsynchronousFetchRequest(fetchRequest: request) { result in
       if let error = result.operationError {
         completion(.failure(error))
@@ -456,7 +482,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
       }
     }
     asynchronousRequest.estimatedResultCount = estimatedResultCount
-
+    
     // swiftlint:disable:next force_cast
     return try context.execute(asynchronousRequest) as! NSAsynchronousFetchResult<Self>
   }